# Add set functions with better performance