Operate in place to save copies and temporary space.

```
# Benchmark test comparing "in place" and "not in place"
$ go test -run=^$ -bench=BenchmarkOperateInPlace -test.benchmem -count=1
os: darwin
goarch: amd64
pkg: istio.io/istio/pkg/util/sets
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
BenchmarkOperateInPlace/Difference-12         	  215138	      5794 ns/op	    1522 B/op	      12 allocs/op
BenchmarkOperateInPlace/DifferenceInPlace-12  	  909678	      1351 ns/op	       0 B/op	       0 allocs/op
BenchmarkOperateInPlace/Intersection-12       	  860622	      1391 ns/op	      48 B/op	       1 allocs/op
BenchmarkOperateInPlace/IntersectInPlace-12   	361817298	         3.302 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	istio.io/istio/pkg/util/sets	6.857s
```