@istio-testing 2024-03-29
@wulianglongrd: The following tests **failed**, say `/retest` to rerun all failed tests or `/retest-required` to rerun all mandatory failed tests:

Test name | Commit | Details | Required | Rerun command
--- | --- | --- | --- | ---
gencheck_istio | 19ab0a1c36c33aa78eab98f96557b9ded2558f1d | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/50158/gencheck_istio/1773605904306933760) | true | `/test gencheck`
lint_istio | 19ab0a1c36c33aa78eab98f96557b9ded2558f1d | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/50158/lint_istio/1773605914289377280) | true | `/test lint`

<details>

Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes/test-infra](https://github.com/kubernetes/test-infra/issues/new?title=Prow%20issue:) repository. I understand the commands that are listed [here](https://go.k8s.io/bot-commands).
</details>
<!-- test report -->